+++ src/main/java/crud/service/UserService.java
@@ -6,8 +6,8 @@ import java.util.List;
 
 public interface UserService {
     List<User> getAllUsers();
-    User findUserById(int id);
+    User findUserById(Long id);
     void saveUser(User user);
     void updateUser(User userToBeUpdated);
-    void removeUserById(int id);
+    void removeUserById(Long id);
 }
